+++ Source/DesignSystem/helpers/ReactRouter/index.tsx
@@ -2,4 +2,3 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 export * from './ReactRouter';
-export * from './DummyLinks';
